add formation direct
select multiple troops and draw a rectangle, they will space themselves
out evenly in the rectangle. this is a bit glitchy. sometimes crashes
the app, not sure why yet.
Sometimes not all the selected troops move. this is because the for
statement stops preemptively. need to find a way to set the rows and
columns variables to overestimate the number of troops selected so that
rows * columns is always greater than noOfTroops