amend formation direct
this was a bit glitchy, sometimes crashed, sometimes didn't move all the
selected troops,
was mostly due to a misplaced bracket causing the approximate distance
between troops value to be inaccurate, which resulted in teh wrong
number of rows and columns which resulted in trying to move more troops
than exist. it doesn't crash anymore but sometimes not all troops move.
after calculating the approximate distance between troops and the
approximate number of columns and rows to distribute the troops to,
there is more logic to calculate how good the approximation was and
whether we need to add 1 or more extra rows or columns, this might be
where the problem, regarding not all troops moving, lies. so fix that
next