camera pan and zoom
this was a lot more complicated than i thought, numerous changes, camera
now successfully pans and zooms, previously it would jump around when
you tried to manipulate camera. only applied camera transformations on
the paint position of the troops. It's supposed to zoom in around the
zoom origin but it doesn't seem to take zoom origin into account at all,
i'll work on that next